# Conflicts:
#	app/src/main/java/ru/netcracker/bikepacker/view/MainNavigationActivity.kt
#	app/src/main/java/ru/netcracker/bikepacker/view/UserMenuInformationAccountFragment.java
#	app/src/main/res/layout/fragment_user_menu_information_account.xml
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/ru/netcracker/bikepacker/adapter/homemenu/HomeMenuRecyclerAdapter.java
@@ -48,7 +48,6 @@ public class HomeMenuRecyclerAdapter extends RecyclerView.Adapter<HomeMenuRecycl
         this.retrofitManager = RetrofitManager.getInstance(context);
         this.userAccountManager = UserAccountManager.getInstance(context);
         this.imageConverter = new ImageConverter();
-
     }
 
     @NonNull
@@ -85,7 +84,7 @@ public class HomeMenuRecyclerAdapter extends RecyclerView.Adapter<HomeMenuRecycl
         holder.bodyDescriptionTextView.setText(newsCardModel.getDescription());
         holder.bodyDescriptionTextView.setOnClickListener(v -> holder.bodyDescriptionTextView.setMaxLines(Integer.MAX_VALUE));
         holder.distanceTextViewNewspaperCard.setText("12");
-        holder.complexityTextViewNewspaperCard.setText(Long.toString(track.getTrackComplexity()));
+        holder.complexityTextViewNewspaperCard.setText(Double.toString(track.getTrackComplexity()));
 
         holder.timeTextViewNewspaperCard.setText(convertTravelTime(track));
         holder.dateHomeMenuItem.setText(newsCardModel.getDate());
